requirements.txt
anthropic>=0.42.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
sentence-transformers>=3.0.0
